[query] enable no implicit numeric widenings warning (#15032)
## Change Description

Enable warnings for implicit numeric widening, e.g. using an Int where a
Long is expected. I like how this makes it completely unambiguous what
numeric types are being used where.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP

+++ hail/hail/src/is/hail/annotations/RegionMemory.scala
@@ -247,7 +247,7 @@ final class RegionMemory(pool: RegionPool) extends AutoCloseable {
   def takeOwnershipOf(r: RegionMemory): Unit =
     references.add(r)
 
-  def nReferencedRegions(): Long = references.size
+  def nReferencedRegions(): Int = references.size
 
   def setNumParents(n: Int): Unit = {
     assert(n >= references.size)
