+++ hail/hail/src/is/hail/expr/ir/MatrixWriter.scala
@@ -2507,7 +2507,7 @@ case class MatrixBlockMatrixWriter(
     val flatPaths = ToArray(mapIR(ToStream(sortedColMajorPairs))(GetTupleElement(_, 1)))
     val bmt = BlockMatrixType(
       elementType,
-      IndexedSeq(numRows, numCols),
+      IndexedSeq(numRows, numCols.toLong),
       numRows == 1,
       blockSize,
       BlockMatrixSparsity.dense,
