+++ hail/hail/test/src/is/hail/expr/ir/TableIRSuite.scala
@@ -809,7 +809,7 @@ class TableIRSuite extends HailSuite {
                 Literal(t, initialData),
                 Some(howManyInitialPartitions),
               ),
-              howManyRowsToTake,
+              howManyRowsToTake.toLong,
             )
           ),
           headData,
@@ -845,7 +845,7 @@ class TableIRSuite extends HailSuite {
                 Literal(t, initialData),
                 Some(howManyInitialPartitions),
               ),
-              howManyRowsToTake,
+              howManyRowsToTake.toLong,
             )
           ),
           headData,
